+++ src/Library/Presentation/Form/DocumentUploadType.php
@@ -4,14 +4,20 @@ declare(strict_types=1);
 
 namespace ChronicleKeeper\Library\Presentation\Form;
 
+use ChronicleKeeper\Library\Application\Service\Document\Importer;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 use Symfony\Component\Form\Extension\Core\Type\FileType;
 use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\Validator\Constraints\File;
 use Symfony\Component\Validator\Constraints\NotNull;
 
 class DocumentUploadType extends AbstractType
 {
+    public function __construct(private readonly Importer $importer)
+    {
+    }
+
     /** @inheritDoc */
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
@@ -21,7 +27,7 @@ class DocumentUploadType extends AbstractType
             [
                 'label' => 'Dokument',
                 'translation_domain' => false,
-                'constraints' => [new NotNull()],
+                'constraints' => [new NotNull(), new File(mimeTypes: $this->importer->getSupportedMimeTypes())],
             ],
         );
         $builder->add(